requirements.txt
# Sushi Catering Management System
# Python 3.x

# Database
SQLAlchemy>=2.0.0

# Discord bot (supports slash commands)
discord.py>=2.3.0

# AI integration (Gemini)
google-genai>=1.0.0

# Environment
python-dotenv>=1.0.0

# Testing
pytest>=7.0.0
